+++ src/api/Weatherstack.hpp
@@ -1,7 +1,7 @@
 /*
  -------------------------------------------------------------------------------
     This file is part of the weather information collector.
-    Copyright (C) 2019  Dirk Stolle
+    Copyright (C) 2019, 2020  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -23,6 +23,7 @@
 
 #include <string>
 #include "API.hpp"
+#include "Plans.hpp"
 
 namespace wic
 {
@@ -32,9 +33,10 @@ class Weatherstack: public API
   public:
     /** \brief Constructor.
      *
+     * \param plan the API pricing plan
      * \param key  the API key for requests
      */
-    Weatherstack(const std::string& key = "");
+    Weatherstack(const PlanWeatherstack plan, const std::string& key = "");
 
 
     /** \brief Sets the API key for API requests.
@@ -117,6 +119,7 @@ class Weatherstack: public API
     #endif // wic_no_json_parsing
   private:
     std::string m_apiKey; /**< the API key for requests */
+    PlanWeatherstack m_plan; /**< the current pricing plan */
 
 
     #ifndef wic_no_network_requests