+++ src/tasks/TaskManager.cpp
@@ -303,7 +303,8 @@ bool TaskManager::loadFromFile(const std::string& fileName, Task& task)
   return true;
 }
 
-bool TaskManager::loadFromDirectory(const std::string& directory, const std::string& extension, std::vector<Task>& storage)
+bool TaskManager::loadFromDirectory(const std::string& directory, const std::string& extension,
+                                    std::vector<Task>& storage, const PlanWeatherstack planWs)
 {
   // Let's keep the namespace short and simple.
   namespace fs = boost::filesystem;
@@ -346,7 +347,7 @@ bool TaskManager::loadFromDirectory(const std::string& directory, const std::str
             return false;
           }
           // Check whether the specified API can handle the task.
-          const auto api = Factory::create(t.api());
+          const auto api = Factory::create(t.api(), planWs);
           if (!api)
           {
             std::cerr << "Error: The API " << toString(t.api())